적 체력 감소 및 플레이어 쫓기
