--- entities/Entity.java
@@ -1,10 +1,0 @@
-package entites;
-
-public abstract class Entity {
-    protected float x,y;
-    public Entity(float x, float y) {
-        this.x = x;
-        this.y = y;
-
-    }
-}
