+++ game/ButtonClickedEvent.java
@@ -1,3 +1,8 @@
+
+
+// 캐릭터 선택 화면에서 필요한 마우스 클릭시 이벤트입니다.
+
+
 package game;
 
 import java.awt.event.MouseAdapter;
