+++ game/CreditPanel.java
@@ -1,3 +1,8 @@
+
+
+// 메인화면에서 '만든이'를 누르거나 게임이 끝날때 나오는 credit 입니다.
+
+
 package game;
 
 public class CreditPanel {
