+++ game/GameFrame.java
@@ -1,3 +1,8 @@
+
+
+// 게임의 전체적인 프레임입니다.
+
+
 package game;
 
 import java.awt.Dimension;
