+++ game/HelpPanel.java
@@ -1,3 +1,8 @@
+
+
+// 시작 화면에서 '도움말' 클릭시 나오는 화면입니다.
+
+
 package game;
 
 public class HelpPanel {
