+++ game/Map2Panel.java
@@ -1,3 +1,8 @@
+
+
+// Map1 클리어 후 이동하는 Map2의 코드입니다.
+
+
 //package game;
 //
 //import entity.Muzan;
