--- main/GameWindow.java
@@ -1,34 +1,0 @@
-package main;
-
-import javax.swing.JFrame;
-import java.awt.event.WindowEvent;
-import java.awt.event.WindowFocusListener;
-
-public class GameWindow {
-    private JFrame jframe;
-    public GameWindow(GamePanel gamePanel) {
-
-        jframe = new JFrame();
-
-        jframe.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        jframe.add(gamePanel);
-        jframe.setLocationRelativeTo(null);
-        jframe.setResizable(false);
-        jframe.pack();
-        jframe.setVisible(true);
-        jframe.addWindowFocusListener(new WindowFocusListener() {
-
-            @Override
-            public void windowLostFocus(WindowEvent e) {
-                gamePanel.getGame().windowFocusLost();
-
-            }
-
-            @Override
-            public void windowGainedFocus(WindowEvent e) {
-
-            }
-        });
-    }
-
-}
