--- main/MainClass.java
@@ -1,7 +1,0 @@
-package main;
-
-public class MainClass {
-    public static void main(String[] args) {
-        new Game();
-    }
-}
