+++ test/levels/Level.java
@@ -1,0 +1,13 @@
+package levels;
+
+public class Level {
+
+    private int[][] lvlData;
+    public Level(int [][] lvlData) {
+        this.lvlData = lvlData;
+    }
+    public int getSpriteIndex(int x, int y) {
+        return lvlData[y][x];
+    }
+
+}
