+++ test/main/MainClass.java
@@ -1,0 +1,7 @@
+package main;
+
+public class MainClass {
+    public static void main(String[] args) {
+        new Game();
+    }
+}
